# libs/rating-service-openapi/openapi.yml
openapi: 3.0.3
info:
  title: NPU Rating API
  description: API for rating NPU creations
  version: 1.0.0


paths:
  /ratings:
    post:
      summary: Submit a rating for an NPU
      operationId: createRating
      tags:
        - Rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingRequest"
      responses:
        "201":
          description: Successfully created a rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResponse"
    get:
      summary: List ratings
      operationId: listRatings
      tags:
        - Rating
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Maximum number of ratings to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
          description: Number of ratings to skip
        - name: npuId
          in: query
          schema:
            type: string
            format: uuid
          description: ID of the NPU to filter ratings by
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: ID of the user to filter ratings by
        - name: scoreGte
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
          description: Score to filter ratings by
        - name: scoreLte
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
          description: Score to filter ratings by
        - name: createdAtGte
          in: query
          schema:
            type: string
            format: date-time
          description: Minimum creation date of the rating
        - name: createdAtLte
          in: query
          schema:
            type: string
            format: date-time
          description: Maximum creation date of the rating
      responses:
        "200":
          description: Returns all ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingResponse"

  /ratings/{id}:
    get:
      summary: Get a rating by ID
      operationId: getRatingById
      tags:
        - Rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Returns the rating details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRatingsResponse"
        "404":
          description: Rating not found

    patch:
      summary: Update an existing rating
      operationId: updateRating
      tags:
        - Rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRatingRequest"
      responses:
        "200":
          description: Successfully updated the rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResponse"
        "404":
          description: Rating not found

    delete:
      summary: Delete a rating by ID
      operationId: deleteRating
      tags:
        - Rating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Rating deleted successfully
        "404":
          description: Rating not found

components:
  schemas:
    CreateRatingRequest:
      type: object
      required: [ npuId, score ]
      properties:
        npuId:
          type: string
          format: uuid
          description: ID of the NPU being rated
        score:
          type: integer
          minimum: 1
          maximum: 10
          description: Score given to the NPU (1-10 scale)
        comment:
          type: string
          description: Free-text feedback from the user

    UpdateRatingRequest:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 10
          description: Updated score (1-10 scale)
        comment:
          type: string
          description: Updated free-text feedback

    RatingResponse:
      type: object
      required: [ id, npuId, userId, score, createdAt, updatedAt ]
      properties:
        id:
          type: number
          description: Unique identifier of the rating
        npuId:
          type: string
          format: uuid
          description: ID of the NPU being rated
        userId:
          type: string
          format: uuid
          description: ID of the user who submitted the rating
        score:
          type: integer
          minimum: 1
          maximum: 10
          description: Score given to the NPU
        comment:
          type: string
          description: Free-text feedback
        createdAt:
          type: string
          format: date-time
          description: Timestamp of when the rating was submitted
        updatedAt:
          type: string
          format: date-time
          description: Timestamp of when the rating was last updated
    ListRatingsResponse:
      type: object
      required: [ pagination, data ]
      properties:
        pagination:
          type: object
          required: [ total, limit, offset ]
          properties:
            total:
              type: number
              description: Total number of ratings
            limit:
              type: number
              description: Maximum number of ratings returned
            offset:
              type: number
              description: Number of ratings to skip
        data:
          type: array
          items:
            $ref: "#/components/schemas/RatingResponse"